requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.1
scikit-learn==1.5.2
PyMuPDF==1.24.9
httpx==0.27.2
mistralai==0.4.2

